+++ src/objects.js
@@ -94,7 +94,7 @@ embedMetadataPNG*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2022-April-25';
+modules.objects = '2022-April-28';
 
 var SpriteMorph;
 var StageMorph;
@@ -1097,6 +1097,12 @@ SpriteMorph.prototype.initBlocks = function () {
             type: 'reporter',
             category: 'sensing',
             spec: '%block of block %repRing',
+            defaults: [['definition']]
+        },
+        doSetBlockAttribute: {
+            type: 'command',
+            category: 'sensing',
+            spec: 'set %byob of block %repRing to %s',
             defaults: [['definition']]
         },
 
@@ -2691,6 +2697,7 @@ SpriteMorph.prototype.blockTemplates = function (
         blocks.push('-');
         blocks.push(block('reportDate'));
         blocks.push(block('reportBlockAttribute'));
+        blocks.push(block('doSetBlockAttribute'));
 
         // for debugging: ///////////////
         if (devMode) {
@@ -9158,6 +9165,7 @@ StageMorph.prototype.blockTemplates = function (
         blocks.push('-');
         blocks.push(block('reportDate'));
         blocks.push(block('reportBlockAttribute'));
+        blocks.push(block('doSetBlockAttribute'));
 
         // for debugging: ///////////////
         if (this.world().isDevMode) {